Patch from Tim Mooney. Text below. ========================================================================== I've discovered another minor "problem" with how RPM handles symlinks. The problem is that rpm calls `realpath()' on all file arguments, so that even if the file I want to check with
	rpm -q -v -f filename

is a symlink that was installed via RPM, what actually happens is that
RPM checks what the symlink points to and reports whether *it* is part of
some package.

Please note that I understand the "Tricky Detail" (page 55 of Maximum RPM)
regarding `-f' and symlinks, but that only relates to symlinks in the path
to the filename, *not* when the filename *is* a symlink.

The problem is in query.c, starting on about line 466.  If `arg' isn't a
full path, realpath() is called on arg.  This means that if `arg' is a symlink,
it's resolved instead of run through rpmFindByFile.

I think what should instead happen is realpath should be called on ".", and
"/" and arg should be appended to that.  That will guarantee that the full
path doesn't have any symlinks, but it will still allow us to do

	rpm -q -v -f filename

on a symlink and have rpm check it, instead of what it points to.

The patch below my sig implements my suggestion.

Note that I see three other areas for improvment in the section of code around
the call to `realpath()'.

1) path is an array of 255 characters.  Instead of using 255, PATH_MAX should
be used if it's defined (a configure test for this could be added, if need
be).

2) the existing code and my new code doesn't boundary check path, so it's
possible to overrun the end of the array.

3) if I feed rpm something like

	rpm -q -v -f wrongfilename

I still get the "not owned by any package", even if the file name I fed it
doesn't exist.  Should an `lstat()' be added in this area and an error message
of

	"file does not exist"

or

	"no such file or directory"

be returned if the lstat fails?

I will implement any of the three additional improvements, if Erik thinks any
of them are a worthwhile change.

Note that there's also a call to realpath() in verify.c, but I didn't look to
see if any of the same changes would be applicable there.

Tim Mooney                              mooney@dogbert.cc.ndsu.NoDak.edu
Information Technology Services         (701) 231-1076 (Voice)
Room 242-J1, IACC Building              (701) 231-8541 (Fax)
North Dakota State University, Fargo, ND 58105-5164