From Tim Mooney:
Make the statfs stuff work more reliably.  Switch from using AC_EGREP_HEADER
to AC_TRY_COMPILE, which should be much more reliable.

Change all places where a previous patch I had sent in said "convertdb"
when it should have said "rpmconvert".

Change the LDFLAGS_STATIC for HP-UX so that static linking isn't done on
HP-UX with libtool.  Everything could be linked statically *except* the shl_*
routines, which are in `libdld.1', but libtool doesn't know how to do
a "--prefer-static" type link, so we have to just avoid trying the static
link at all.

Fix a warning that happens over and over as `header.h' is included
on Digital Unix systems.  Digital Unix defines `__alpha', not `__alpha__',
so I just added the additional check for __alpha.

Fix a number of places in the code where there is a spurious/extra
semi-colon.  This gets rid of quite a few warnings.