Although %setup and %patch look like regular RPM macros on the first spot they are not real macros, of course. Instead they are parsed and treated very special by RPM internally.
Unfortunately, this internal handling didn't allow for any leading
whitespaces in front of %setup and %patch in *.spec files. This is a
problem if one uses ultra strictly formatted .spec files (as we do in
OpenPKG) where all section content is consequently indented.

Hence for cosmetics and consistency addicts like me this short-coming in
the parsing of %setup and %patch was never acceptable, so I applied this
little change to the OpenPKG RPM since six years now. I think it will
not cause trouble, but better to let it be reviewed by others first.