Another "Build Environment Anti-Convolution Effort":
PART 1:

Finally revise the building against the _INTERNAL_ copies of third-party
libraries in order to circumvent the problem of extending CPPFLAGS,
LDFLAGS and LIBS in this particular case:

1. at "configure" time at least the library file (libfoo.a)
   is still not existing, so extending LIBS with "-lfoo" would cause
   following Autoconf checks to fail.

2. even deferring the extension of LIBS doesn't work, because although
   this works around the problem under (1), it will fail if more than
   one internal third-party library exists: LIBS would contain "-lfoo
   -lbar" and on building "foo", "bar" usually still isn't built (or
   vice versa). Hence, the linking of programs (tools, tests, etc) in
   "foo" would fail.

3. information in at least LDFLAGS and LIBS is usually passed-through
   to applications via xxx-config(1) scripts or pkg-config(1)
   specifications. As the path to the internal copy is usually just a
   temporary path, this will break there, too.

So, internal copies of third-party libraries inherently have to be
handled _explicitly_ by the build environment and for this we can only
provide the necessary information in dedicated per-library variables.

PART 2:

Additionally, the code of all internal third-party has to be _installed_
or librpm* would be useless in practice. For Berkeley-DB this is
achieved since a longer time by bundling its object files directly with
librpmdb. But for Lua, File and Zlib the same is required. As librpmdb
is not the best location for all of them and all third-party code
should be bundled together, after lots of thinking I've decided to use
librpmmisc for this purpose. Especially, because librpmmisc is already
a "bare container-style library" and it is at the leave of the library
dependency tree.

Finally, the problematic variable overides like "LIBS =" in a few
Makefile.am were completely eliminated by providing a $(LIBS)
independent linking target for the libraries.

RESULTS:

The resulting RPM libraries now look like this (notice the larger
librpmmisc*: it contains Berkeley-DB, Lua, File and ZLib here):

    -rwxr-xr-x   515781 Jun 27 21:36 librpm-5.0.so
    -rw-r--r--   595346 Jun 27 21:36 librpm.a
    -rwxr-xr-x     1448 Jun 27 21:36 librpm.la
    lrwxr-xr-x       13 Jun 27 21:36 librpm.so -> librpm-5.0.so
    -rwxr-xr-x   141717 Jun 27 21:36 librpmbuild-5.0.so
    -rw-r--r--   165622 Jun 27 21:36 librpmbuild.a
    -rwxr-xr-x     1483 Jun 27 21:36 librpmbuild.la
    lrwxr-xr-x       18 Jun 27 21:36 librpmbuild.so -> librpmbuild-5.0.so
    -rwxr-xr-x   162454 Jun 27 21:36 librpmdb-5.0.so
    -rw-r--r--   172300 Jun 27 21:36 librpmdb.a
    -rwxr-xr-x     1462 Jun 27 21:36 librpmdb.la
    lrwxr-xr-x       15 Jun 27 21:36 librpmdb.so -> librpmdb-5.0.so
    -rwxr-xr-x   302572 Jun 27 21:36 librpmio-5.0.so
    -rw-r--r--   323572 Jun 27 21:36 librpmio.a
    -rwxr-xr-x     1462 Jun 27 21:36 librpmio.la
    lrwxr-xr-x       15 Jun 27 21:36 librpmio.so -> librpmio-5.0.so
    -rwxr-xr-x  1597583 Jun 27 21:36 librpmmisc-5.0.so
    -rw-r--r--  2030858 Jun 27 21:36 librpmmisc.a
    -rwxr-xr-x     1577 Jun 27 21:36 librpmmisc.la
    lrwxr-xr-x       17 Jun 27 21:36 librpmmisc.so -> librpmmisc-5.0.so

This way the convoluted linking against internal and external
third-party libraries was cleaned up. The only real "difference" to the
past is that the installed librpm* libraries no longer directly carry
(via ELF headers or GNU libtool *.la files) their dependencies (at least
the ELF headers cannot be used across platforms anyway). But as we now
provide a standard "rpm.pc" file, RPM based applications neither any
longer need those implicit dependencies nor do they have to know it
theirself. Instead the canonical and easy way to build against the RPM
5 API (while taking care of library ordering and external dependencies)
is:

    CPPFLAGS = `pkg-config rpm --cflags-only-I`
    CFLAGS   = `pkg-config rpm --cflags-only-other`
    LDFLAGS  = `pkg-config rpm --libs-only-other`
    LIBS     = `pkg-config rpm --libs-only-l`