.open "Peril Sensitive Sunglasses" .load "Armor" "Maximum" .strike "Argumentation-Flag" "half-mast"
Based on all the recent cleanup and streamline efforts on the RPM 5
build environment, it is now possible to surgically teach RPM 5 one
remaining (officially unsupported but) technically very interesting
feature:

   Out-of-the-box building RPM without Berkeley-DB at all.

This is explicitly declared "unsupported" in Autoconf by default
and a simple "--without-db" is intentionally _NOT_ sufficient. One
really has to use "--with-bugreport=<email-address> --without-db
--with-sqlite=...", i.e., one has to explicitly set a support channel
different from "rpm-devel@rpm5.org" and one at the same time has to
provide at least the alternative SQLite-based RPM DB backend. So,
the packager really has to take _EXPLICIT_ action and this is NOT
recommended by us although it is now at least technically possible.

In case of an SQLite-only RPM the "macros" file is automatically
adjusted to "_dbapi 4", so this really works out-of-the-box also
under runtime. In case one builds RPM with both Berkeley-DB and
SQLite the default is still Berkeley-DB, of course. Nevertheless one
can now explicitly select the configured default RPM DB API with
--with-dbapi={db,sqlite} now, too.