It is both confusing and problematic that the current RPM 5 on HEAD labels itself as just "5.0" and this way looks like the final release version. As for the snapshots we also need a clear labeling, let's start using "N.M.DEVEL" for the development state in CVS and "M.M.SNAPSHOT.YYYYMMDD" for the daily snapshots from CVS.
This way the snapshot and development versions can be clearly
distinguished and especially not confused with any forthcoming RPM 5
_release_ versions.

BTW, I've thought very long what separation character to use. Other
OSS projects usually use "-" instead of "." in front of "DEVEL" and
"SNAPSHOT", but this is too problematic for us in the RPM world (think
about the usual <name>-<version>-<release> we have). Hence, I stick with
a dot for now as this seems to be the best compromise.