Remove the extra check as RPM already does what OpenPKG wants: only tags which are explicitly listed in %{_arbitrary_tags} are allowed. I thought the initial intention of RPM was to allow really even *unknown* tags, but what is actually implemented is different as tests showed...
Independent how RPM is further hacked on: we really need a way to still
allow only tags which are explicitly listed. Whether this is the default
(as it is the case currently) or has to be achieved via some special
configuration does not matter. Please keep this in mind if you want to
change RPM to accept really *unknown* tags...