build/files.c (processBinaryFiles): break early on errors
They say "when you must fail, fail noisily and as soon as possible".
So I change processBinaryFiles() as follows:

1) When processing a package, if a fatal error occurs, I simply
break the loop with RPMRC_FAIL.  Other stages for the current
package are skipped, and other sub-packages are simply not processed
at all.  (This is actually how older rpm versions behaved.)

2) If a fatal error occurs, extra checks are skipped.  This now
includes only checkFiles() check for unpackaged files.  (More
checks to be implemented.)