During the recalculation of the buildid, it's necessary to change the word back to the original endian.  However, if we do this in-place, we've also affected the headers that we're also working on.  The side effect of this is we can no longer rely on 'sh_type' as it may have been changed.
This patch ensures that any time we translate the loaded data to the machine
format, we only do it in a backup copy and never the original copy.

Note: in all other places a backup copy was used, just not buildid processing.

Also the process (...) function was modified to verify the data is not
NULL as well.  This is an extra check and is not strictly necessary.

Upstream-Status: Pending

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>