rpm-python-module: Change the extension tag from PyCObject to PyInt
Use the tagValue to determine the custom PyInt value to use for the extension
tag.  Without this, any custom tag extensions will be returned in a format
that the tagNumFromPyObject and related functions like hdr_subscript will
failed to process.  Usually the failure is error: expected a string or integer

Upstream-Status: Pending

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>